extract page header into a separate component
